+++ AmplifyPlugins/API/Tests/AWSAPIPluginTests/Operation/GraphQLSubscribeTaskTests.swift
@@ -11,7 +11,7 @@ import XCTest
 import Amplify
 @testable import AmplifyTestCommon
 @testable import AWSAPIPlugin
-import AmplifyAsyncTesting
+@_implementationOnly import AmplifyAsyncTesting
 import AppSyncRealTimeClient
 
 class GraphQLSubscribeTasksTests: OperationTestBase {
