+++ AmplifyPlugins/Storage/Tests/AWSS3StoragePluginTests/Support/Internal/StorageTransferDatabaseTests.swift
@@ -10,7 +10,7 @@ import XCTest
 @testable import AWSS3StoragePlugin
 @testable import Amplify
 import AmplifyTestCommon
-import AmplifyAsyncTesting
+@_implementationOnly import AmplifyAsyncTesting
 
 // swiftlint:disable line_length
